+++ include/title_id.h
@@ -1,5 +1,5 @@
-#ifndef _ISO_TITLE_ID_H_
-#define _ISO_TITLE_ID_H_
+#ifndef _TITLE_ID_H_
+#define _TITLE_ID_H_
 
 // Loads SYSTEM.CNF from ISO and extracts title ID
 char *getTitleID(char *path);
