--- src/iso_cache.c
@@ -1,255 +1,0 @@
-// Implements title ID cache for file-based devices to make building target list faster
-#include "iso_cache.h"
-#include "common.h"
-#include "devices.h"
-#include "target.h"
-#include "options.h"
-#include <malloc.h>
-#include <ps2sdkapi.h>
-#include <stdio.h>
-#include <string.h>
-
-#define CACHE_MAGIC "NIDC"
-#define CACHE_VERSION 2
-
-const char titleIDCacheFile[] = "/cache.bin";
-
-// Structs used to read and write cache file contents
-typedef struct {
-  char titleID[12];
-  size_t pathLength; // Includes null-terminator
-} CacheEntryHeader;
-
-typedef struct {
-  char magic[4];   // Must be always equal to CACHE_MAGIC
-  uint8_t version; // Cache version
-  int total;       // Total number of elements in cache file
-} CacheMetadata;
-
-// Saves TargetList into title ID cache on given storage device
-int storeTitleIDCache(TargetList *list, struct DeviceMapEntry *device) {
-  if (list->total == 0) {
-    return 0;
-  }
-
-  // Get total number of valid cache entries
-  int total = 0;
-  Target *curTitle = list->first;
-  while (curTitle != NULL) {
-    if (strlen(curTitle->id) == 11) {
-      total++;
-    }
-    curTitle = curTitle->next;
-  }
-  if (total == 0) {
-    printf("WARN: No valid cache entries found\n");
-    return 0;
-  }
-
-  // Make sure path exists
-  if (device->mode == MODE_NONE || device->mountpoint == NULL)
-    return -ENODEV;
-
-  // Prepare paths and header
-  char cachePath[PATH_MAX];
-  char dirPath[PATH_MAX];
-  CacheEntryHeader header;
-  CacheMetadata meta = {.magic = CACHE_MAGIC, .version = CACHE_VERSION, .total = total};
-
-  buildConfigFilePath(dirPath, device->mountpoint, NULL);
-  buildConfigFilePath(cachePath, device->mountpoint, titleIDCacheFile);
-
-  // Get path to config directory and make sure it exists
-  struct stat st;
-  if (stat(dirPath, &st) == -1) {
-    printf("Creating config directory: %s\n", dirPath);
-    if (mkdir(dirPath, 0777)) {
-      printf("ERROR: Failed to create directory\n");
-      return -EIO;
-    }
-  }
-
-  // Open cache file for writing
-  FILE *file = fopen(cachePath, "wb");
-  if (file == NULL) {
-    printf("ERROR: Failed to open cache file for writing\n");
-    return -EIO;
-  }
-
-  int result;
-  // Write cache file header
-  result = fwrite(&meta, sizeof(CacheMetadata), 1, file);
-  if (!result) {
-    printf("ERROR: Failed to write metadata: %d\n", errno);
-    fclose(file);
-    remove(cachePath);
-    return result;
-  }
-
-  // Write each entry
-  curTitle = list->first;
-  int mountpointLen = -1;
-  while (curTitle != NULL) {
-    // Ignore empty entries or entries not belonging to the current device
-    if ((strlen(curTitle->id) < 11) || (curTitle->device != device)) {
-      curTitle = curTitle->next;
-      continue;
-    }
-
-    // Compare paths without the mountpoint
-    mountpointLen = getRelativePathIdx(curTitle->fullPath);
-    if (mountpointLen == -1) {
-      printf("WARN: Failed to get device mountpoint for %s\n", curTitle->name);
-      curTitle = curTitle->next;
-      continue;
-    }
-
-    // Write entry header
-    memcpy(header.titleID, curTitle->id, sizeof(header.titleID));
-    header.titleID[11] = '\0';
-    header.pathLength = strlen(curTitle->fullPath) - mountpointLen + 1;
-    result = fwrite(&header, sizeof(CacheEntryHeader), 1, file);
-    if (!result) {
-      printf("ERROR: %s: Failed to write header: %d\n", curTitle->name, errno);
-      fclose(file);
-      remove(cachePath);
-      return result;
-    }
-    // Write full ISO path without the mountpoint
-    result = fwrite(curTitle->fullPath + mountpointLen, header.pathLength, 1, file);
-    if (!result) {
-      printf("ERROR: %s: Failed to write full path: %d\n", curTitle->name, errno);
-      fclose(file);
-      remove(cachePath);
-      return result;
-    }
-    curTitle = curTitle->next;
-  }
-  fclose(file);
-
-  return 0;
-}
-
-// Loads title ID cache from storage into cache
-int loadTitleIDCache(TitleIDCache *cache, struct DeviceMapEntry *device) {
-  // Make sure path exists
-  if (device->mode == MODE_NONE || device->mountpoint == NULL)
-    return -ENODEV;
-
-  cache->total = 0;
-  cache->lastMatchedIdx = 0;
-
-  // Open cache file for reading
-  char cachePath[PATH_MAX];
-
-  FILE *file;
-  // Load the first found cache file
-  buildConfigFilePath(cachePath, device->mountpoint, titleIDCacheFile);
-
-  file = fopen(cachePath, "rb");
-  if (file == NULL)
-    return -ENOENT;
-
-  int result;
-
-  // Read cache file header
-  CacheMetadata meta;
-  result = fread(&meta, sizeof(CacheMetadata), 1, file);
-  if (!result) {
-    printf("ERROR: Failed to read cache metadata\n");
-    fclose(file);
-    return result;
-  }
-
-  // Make sure header is valid
-  if (!strcmp(meta.magic, CACHE_MAGIC)) {
-    printf("ERROR: Cache magic doesn't match, refusing to load\n");
-    fclose(file);
-    return -EINVAL;
-  }
-  if (meta.version != CACHE_VERSION) {
-    printf("ERROR: Unsupported cache version %d\n", meta.version);
-    fclose(file);
-    return -EINVAL;
-  }
-
-  // Allocate memory for cache entries based on total entry count from header metadata
-  int readIndex = 0;
-  cache->entries = malloc((sizeof(CacheEntry) * meta.total));
-  if (cache->entries == NULL) {
-    printf("ERROR: Can't allocate enough memory\n");
-    fclose(file);
-    return -ENOMEM;
-  }
-
-  // Read each entry into cache
-  CacheEntryHeader header;
-  char pathBuf[PATH_MAX + 1];
-  while (!feof(file)) {
-    // Read cache entry header
-    pathBuf[0] = '\0';
-    result = fread(&header, sizeof(CacheEntryHeader), 1, file);
-    if (result != 1) {
-      if (!feof(file))
-        printf("WARN: Read less than expected, title ID cache might be incomplete\n");
-      break;
-    }
-    // Read ISO path
-    result = fread(&pathBuf, header.pathLength, 1, file);
-    if (result != 1) {
-      printf("WARN: Read less than expected, title ID cache might be incomplete\n");
-      break;
-    }
-    pathBuf[header.pathLength] = '\0';
-
-    // Store entry in cache
-    CacheEntry entry;
-    memcpy(entry.titleID, header.titleID, sizeof(entry.titleID));
-    header.titleID[11] = '\0';
-    entry.fullPath = strdup(pathBuf);
-    cache->entries[readIndex] = entry;
-    readIndex++;
-  }
-  fclose(file);
-
-  // Free unused memory
-  if (readIndex != meta.total)
-    cache->entries = realloc(cache->entries, sizeof(CacheEntry) * readIndex);
-
-  cache->total = readIndex;
-  return 0;
-}
-
-// Returns a pointer to title ID or NULL if fullPath is not found in the cache
-char *getCachedTitleID(char *fullPath, TitleIDCache *cache) {
-  // This code takes advantage of all entries in the title list being sorted alphabetically.
-  // By starting from the index of the last matched entry, we can skip comparing fullPath with entries
-  // that have already been matched to a title ID, improving lookup speeds for very large lists.
-  int mountpointLen = getRelativePathIdx(fullPath);
-  if (mountpointLen == -1) {
-    printf("WARN: Failed to get device mountpoint for %s\n", fullPath);
-    return NULL;
-  }
-
-  for (int i = cache->lastMatchedIdx; i < cache->total; i++) {
-    if (!strcmp(cache->entries[i].fullPath, fullPath + mountpointLen)) {
-      cache->lastMatchedIdx = i;
-      return cache->entries[i].titleID;
-    }
-  }
-  return NULL;
-}
-
-// Frees memory used by title ID cache
-// All pointers to cache entries (including title IDs) will be invalid
-void freeTitleCache(TitleIDCache *cache) {
-  if (cache == NULL)
-    return;
-
-  for (int i = 0; i < cache->total; i++) {
-    free(cache->entries[i].fullPath);
-  }
-
-  free(cache->entries);
-  free(cache);
-}
